separated common util functions to eliminate duplicities of code and have all build mechanism include new files
