+++ contrib/QtCreator/MEGAcmd/MEGAcmdServer/MEGAcmdServer.pro
@@ -61,7 +61,8 @@ SOURCES += ../../../../src/megacmd.cpp \
     ../../../../src/megacmdsandbox.cpp \
     ../../../../src/configurationmanager.cpp \
     ../../../../src/comunicationsmanager.cpp \
-    ../../../../src/megacmdutils.cpp
+    ../../../../src/megacmdutils.cpp \
+    ../../../../src/megacmdcommonutils.cpp
 
 
 HEADERS += ../../../../src/megacmd.h \
@@ -72,6 +73,7 @@ HEADERS += ../../../../src/megacmd.h \
     ../../../../src/configurationmanager.h \
     ../../../../src/comunicationsmanager.h \
     ../../../../src/megacmdutils.h \
+    ../../../../src/megacmdcommonutils.h \
     ../../../../src/megacmdversion.h \
     ../../../../src/megacmdplatform.h
