+++ src/comunicationsmanagernamedpipes.cpp
@@ -109,7 +109,7 @@ HANDLE ComunicationsManagerNamedPipes::create_new_namedPipe(int *pipeId)
             {
                 LOG_fatal << "ERROR opening namedPipe ID=" << pipeId << " errno: " << ERRNO << ". Attempts: " << attempts;
             }
-            sleepMicroSeconds(500);
+            sleepMilliSeconds(500);
         }
         else
         {
@@ -183,7 +183,7 @@ int ComunicationsManagerNamedPipes::waitForPetition()
         else
         {
             LOG_fatal << "ERROR on connecting to namedPipe. errno: " << ERRNO;
-            sleepMicroSeconds(1000);
+            sleepMilliSeconds(1000);
             pipeGeneral = INVALID_HANDLE_VALUE;
             return false;
         }
@@ -247,7 +247,7 @@ void ComunicationsManagerNamedPipes::returnAndClosePetition(CmdPetition *inf, OU
             {
                 LOG_fatal << "ERROR on connecting to namedPipe " << outNamedPipe << ". errno: " << ERRNO << ". Attempts: " << attempts;
             }
-            sleepMicroSeconds(500);
+            sleepMilliSeconds(500);
         }
         else
         {
@@ -302,7 +302,7 @@ void ComunicationsManagerNamedPipes::sendPartialOutput(CmdPetition *inf, OUTSTRI
             {
                 cerr << "ERROR on connecting to namedPipe " << outNamedPipe << ". errno: " << ERRNO << ". Attempts: " << attempts << endl;
             }
-            sleepMicroSeconds(500);
+            sleepMilliSeconds(500);
         }
         else
         {
@@ -370,7 +370,7 @@ void ComunicationsManagerNamedPipes::sendPartialOutput(CmdPetition *inf, char *s
             {
                 cerr << "ERROR on connecting to namedPipe " << outNamedPipe << ". errno: " << ERRNO << ". Attempts: " << attempts << endl;
             }
-            sleepMicroSeconds(500);
+            sleepMilliSeconds(500);
         }
         else
         {
