+++ src/megacmdshell/megacmdshellcommunications.h
@@ -22,6 +22,8 @@
 #ifndef MEGACMDSHELLCOMMUNICATIONS_H
 #define MEGACMDSHELLCOMMUNICATIONS_H
 
+#include "../megacmdcommonutils.h"
+
 #include <string>
 #include <iostream>
 
@@ -51,25 +53,10 @@ class MegaThread : public ::mega::PosixThread {};
 #endif
 
 #ifdef _WIN32
-
-#define OUTSTREAMTYPE std::wostream
-#define OUTSTRINGSTREAM std::wostringstream
-#define OUTSTRING std::wstring
-#define COUT std::wcout
-
-std::wostream & operator<< ( std::wostream & ostr, std::string const & str );
-std::wostream & operator<< ( std::wostream & ostr, const char * str );
-std::ostringstream & operator<< ( std::ostringstream & ostr, std::wstring const &str);
-
 #else
-#define OUTSTREAMTYPE std::ostream
-#define OUTSTRINGSTREAM std::ostringstream
-#define OUTSTRING std::string
-#define COUT std::cout
 typedef int SOCKET;
 #endif
 
-#define OUTSTREAM COUT
 
 #ifdef _WIN32
 void stringtolocalw(const char* path, std::wstring* local);
